requirements.txt
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.31
langchain-ollama==0.0.10
pinecone-client
python-dotenv==1.0.1
numpy==1.26.4
faiss-cpu==1.7.4  # Use faiss-gpu if you have CUDA support
warnings==0.1.0